requirements.txt
# AI SERVICES
groq==0.6.0
langchain==0.2.14
langchain-community==0.2.12
langchain-anthropic==0.1.22
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-openai==0.1.22
openai==1.51.2
fastapi==0.109.2

# DATA PROCESSING AND ANALYSIS
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.4
xlsxwriter

# WEB AND API
requests==2.25.1
httpx==0.27.0
bs4==0.0.2

# GUI AND VISUALIZATION
PySide6==6.4.2
customtkinter==5.2.2
streamlit==1.35.0
NodeGraphQt==0.6.37

# DEVELOPMENT TOOLS
python-dotenv
black==24.10.0
astor==0.8.1
radon==6.0.1
icecream==2.1.3

# TEXT PROCESSING
markdown-it-py
markdown==3.7
docx==0.2.4
pygments==2.18.0
tiktoken==0.7.0

# UTILITIES
asyncio==3.4.3
pyautogui==0.9.54
pydantic==2.6.1
pygetwindow==0.0.9
pyperclip==1.9.0
pytesseract==0.3.10
tkinterdnd2==0.3.0
toml==0.10.2
colorama

watchdog
rich==13.3.4
